% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoapp_queryDB.R
\name{demoapp_queryDB}
\alias{demoapp_queryDB}
\title{Query SQLite Database for Data}
\usage{
demoapp_queryDB(selected_table = "", database_dir = "", columns)
}
\arguments{
\item{selected_table}{The table inside of the Database the user wishes to query}

\item{database_dir}{A string with the path directory to the database}

\item{columns}{For if the user wishes to only query select columns, the default retrieves all columns or the user can select \code{columns = "all"}, but a character vector can be supplied like so: \code{columns = c("text", "sentiment", "permalink")} for being selective.}
}
\value{
The queried table as a data frame with selected columns if provided
}
\description{
Query SQLite Database for Data
}
\examples{
automotive_df_queried <- demoapp_queryDB(selected_table = "automotive_df",
 database_dir = "demoApp_db.sqlite")

automotive_df_queried <- demoapp_queryDB(selected_table = "automotive_df",
 database_dir = "demoApp_db.sqlite",
 columns = c("text", "sentiment", "permalink"))
}
