% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoapp_writetoDB.R
\name{demoapp_writetoDB}
\alias{demoapp_writetoDB}
\title{Connect and Write Files to a SQLite Database}
\usage{
demoapp_writetoDB(
  data,
  name = "",
  database_path = NULL,
  overwrite = TRUE,
  append = FALSE
)
}
\arguments{
\item{data}{Data frame object that you wish to be a table in the database}

\item{name}{A string representing the name of the table that the user wishes to write}

\item{database_path}{A string with the path directory to the database, the default is set to whatever the current working directory is using \code{getwd()}, searching for \code{database.sqlite} in that project directory. Note: If the database does not yet exist, the function will write it as well as the file, and inform the user within the R console.}

\item{overwrite}{A logical argument as to whether the user wishes to overwrite any existing data with the same name}

\item{append}{A logical argument as to whether the user wishes to append new data on existing data with the same name}
}
\value{
Writes a retrievable data table to a SQLite database
}
\description{
Connect and Write Files to a SQLite Database
}
\examples{
demoapp_writetoDB(data = df_automotive,
 name = "automotive_df", 
 overwrite = TRUE,
  append = FALSE)
}
