% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{run_app}
\usage{
run_app(port = NULL)
}
\arguments{
\item{port}{The TCP port that the application should listen on
if NULL, a random one is assigned}
}
\value{
NA
}
\description{
run_app
}
