% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDemoApp.R
\name{runDemoApp}
\alias{runDemoApp}
\title{Run the demo application}
\usage{
runDemoApp(port = 3838, ...)
}
\arguments{
\item{port}{port to use for webserver}

\item{...}{additional arguments to \code{\link[shiny]{runApp}}}
}
\value{
nothing
}
\description{
Run the demo application
}
