% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoapp_writetoDB.R
\name{demoapp_writetoDB}
\alias{demoapp_writetoDB}
\title{Write Files to a SQLite Database}
\usage{
demoapp_writetoDB(
  data,
  name = "",
  database_dir = "",
  overwrite = FALSE,
  append = FALSE
)
}
\arguments{
\item{data}{Data frame object that you wish to be a table in the database}

\item{name}{A string representing the name of the table that the user wishes to write}

\item{database_dir}{A string with the path directory to the database}

\item{overwrite}{A logical argument as to whether the user wishes to overwrite any existing data with the same name}

\item{append}{A logical argument as to whether the user wishes to append new data on existing data with the same name}
}
\value{
A retrievable data table in the desired database
}
\description{
Write Files to a SQLite Database
}
\examples{
demoapp_writetoDB(data = df_automotive,
name = "automotive_df",
database_dir = "demoApp_db.sqlite",
overwrite = FALSE,
append = FALSE)
}
